requirements.txt
aiohttp==3.7.4
cachetools==4.2.2
colorama==0.4.4
git+https://github.com/Rapptz/discord.py.git@1e17b7fceaf7173b1666465e639d1bebaa126683
git+https://github.com/readthedocs/sphinx-hoverxref@master
python-dotenv==0.12.0
pychalk==2.0.1
pymongo[srv]==3.11.4
Pillow==8.2.0
readthedocs-sphinx-search==0.1.0
Sphinx==4.1.2
sphinx-rtd-dark-mode==1.2.4
sphinx-rtd-theme==0.5.2
sphinx-tabs==3.2.0
topggpy==1.3.0